requirements.txt
appdirs==1.4.3
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
certifi==2017.4.17
cssselect==0.9.1
dj-places==1.1.1
Django==1.10.7
django-filer==1.2.7
django-fontawesome==0.3.1
django-grappelli==2.9.1
django-libs==1.67.12
django-mptt==0.8.7
django-polymorphic==1.0.2
django-suit==0.2.25
easy-thumbnails==2.3
enum34==1.1.6
feedfinder2==0.0.1
feedparser==5.1.3
geograpy==0.3.7
idna==2.5
jellyfish==0.5.6
jieba==0.35
langdetect==1.0.7
lxml==3.3.5
newspaper==0.0.9.8
nltk==2.0.5
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==2.5.1
pycountry==17.1.8
pyparsing==2.2.0
python-dateutil==2.4.0
pytz==2017.2
PyYAML==3.11
requests==2.3.0
requests-file==1.4.1
satori-sdk-python==1.0.3
six==1.7.3
tldextract==1.5.1
Unidecode==0.4.20
vaderSentiment==2.5
